requirements.txt
django==4.2
requests
djangorestframework
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers
numpy==1.25.2
pandas==2.0.1